+++ src/lib_rules.ml
@@ -162,11 +162,11 @@ module Gen (P : Install_rules.Params) = struct
     let cctx = Compilation_context.for_wrapped_compat cctx wrapped_compat in
     Module_compilation.build_modules cctx ~js_of_ocaml ~dynlink ~dep_graphs
 
-  let build_c_file (lib : Library.t) ~expander ~dir ~includes (loc, src, dst) =
+  let build_c_file (lib : Library.t) ~dir ~expander ~includes (loc, src, dst) =
     let src = C.Source.path src in
+    let c_flags = SC.c_flags sctx ~dir ~expander ~lib in
     SC.add_rule sctx ~loc ~dir
-      (Expander.expand_and_eval_set expander lib.c_flags
-         ~standard:(Build.return (Context.cc_g ctx))
+      (c_flags
        >>>
        Build.run
          (* We have to execute the rule in the library directory as
@@ -181,7 +181,7 @@ module Gen (P : Install_rules.Params) = struct
          ]);
     dst
 
-  let build_cxx_file (lib : Library.t) ~expander ~dir ~includes (loc, src, dst) =
+  let build_cxx_file (lib : Library.t) ~dir ~expander ~includes (loc, src, dst) =
     let src = C.Source.path src in
     let open Arg_spec in
     let output_param =
@@ -190,9 +190,9 @@ module Gen (P : Install_rules.Params) = struct
       else
         [A "-o"; Target dst]
     in
+    let cxx_flags = SC.cxx_flags sctx ~dir ~expander ~lib in
     SC.add_rule sctx ~loc ~dir
-      (Expander.expand_and_eval_set expander lib.cxx_flags
-         ~standard:(Build.return (Context.cc_g ctx))
+      (cxx_flags
        >>>
        Build.run
          (* We have to execute the rule in the library directory as
@@ -200,7 +200,6 @@ module Gen (P : Install_rules.Params) = struct
          ~dir
          (SC.resolve_program ~loc:None ~dir sctx ctx.c_compiler)
          ([ S [A "-I"; Path ctx.stdlib_dir]
-          ; As (SC.cxx_flags sctx)
           ; includes
           ; Dyn (fun cxx_flags -> As cxx_flags)
           ] @ output_param @
@@ -281,7 +280,7 @@ module Gen (P : Install_rules.Params) = struct
       String.Map.to_list files
       |> List.map ~f:(fun (obj, (loc, src)) ->
         let dst = Path.relative dir (obj ^ ctx.ext_obj) in
-        build_x lib ~expander ~dir ~includes (loc, src, dst)
+        build_x lib ~dir ~expander ~includes (loc, src, dst)
       )
     in
     let { C.Kind.Dict. c; cxx } = C.Sources.split_by_kind c_sources in