+++ src/pform.ml
@@ -9,6 +9,8 @@ module Var = struct
     | Deps
     | Targets
     | Named_local
+    | Cc
+    | Cxx
 
   let to_sexp =
     let open Sexp.Encoder in
@@ -19,6 +21,8 @@ module Var = struct
     | Deps -> string "Deps"
     | Targets -> string "Targets"
     | Named_local -> string "Named_local"
+    | Cc -> string "cc"
+    | Cxx -> string "cxx"
 
   let pp_debug fmt t =
     Sexp.pp fmt (to_sexp t)
@@ -154,7 +158,7 @@ module Map = struct
       ; "env", since ~version:(1, 4) Macro.Env
       ]
 
-  let create ~(context : Context.t) ~cxx_flags =
+  let create ~(context : Context.t) =
     let ocamlopt =
       match context.ocamlopt with
       | None -> Path.relative context.ocaml_bin "ocamlopt"
@@ -168,6 +172,8 @@ module Map = struct
       | Some p -> path p
     in
     let cflags = context.ocamlc_cflags in
+    let cxx_flags = List.filter context.ocamlc_cflags
+                      ~f:(fun s -> not (String.is_prefix s ~prefix:"-std=")) in
     let strings s = values (Value.L.strings s) in
     let lowercased =
       [ "cpp"            , strings (context.c_compiler :: cflags @ ["-E"])