+++ src/pform.mli
@@ -8,6 +8,8 @@ module Var : sig
     | Deps
     | Targets
     | Named_local
+    | Cc
+    | Cxx
 end
 
 module Macro : sig
@@ -38,7 +40,7 @@ end
 module Map : sig
   type t
 
-  val create : context:Context.t -> cxx_flags:string list -> t
+  val create : context:Context.t -> t
 
   val superpose : t -> t -> t