+++ test/blackbox-tests/test-cases/env/printenv.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+dune printenv $@