+++ test/blackbox-tests/test-cases/workspaces/run.t
@@ -65,4 +65,6 @@ Workspaces also allow you to set the env for a context:
    (flags (-w -40 -machin))
    (ocamlc_flags (-g -verbose))
    (ocamlopt_flags (-g))
+   (c_flags ())
+   (cxx_flags ())
   )